[MVP] general use of pkl
DataSourceFile now use pkl when creating a table:

When the user create the source or update it a data.pkl file will be
created. (in addition to the data.csv/xlsx/...)
When the user create a table from this source, the pickle file is used

This greatly increase the speed, however, there is a drawback. If the
user change the file manually directly in the project folder the
data.pkl file won't be updated.

A solution could be to implement the sync method for the DataSourceFile?