[MVP] refactor UT: fix perf issue
When creating a project, submit on pop up redirect on another page.
We checked that pop up visibility was hidden, but has pop up did not
exist anymore in the new page, this took few seconds each time.

Fix this by adding a new parameter 'redirect' that bypass the check on
visibility