[MVP] imp Pipeline class
_load_actions should only be called in init, as every other actions are
called after the object is initiated. This improve perfs (no need to
read the file multiple times) but this is also more robust. Indeed, if
the object is updated but not saved, than we call another function, this
other function will take the actions not updated yet, this could lead to
bugs difficult to debug.

+ misc imp and remove one useless TODO point