[MVP] mp error management when error in pipeline
when an error was raised when running the pipeline we were redirected to
the project error page instead of the table one. Because of that it was
impossible to go on pipeline's page an edit the failing action. This
commit solves that.

Additionally it introduces a new the error field inside the pipeline
action model. It enables to know what action failed at last run and for
what reason